switching to ingest decimal prices, and to utilise a websocket approach

+++ backend/src/main/java/tradeMatcher/LevelInfo.java
@@ -1,15 +1,15 @@
 package tradeMatcher;
 
 public final class LevelInfo {
-    private final int price;
+    private final double price;
     private final int quantity;
 
-    public LevelInfo(int price, int quantity) {
+    public LevelInfo(double price, int quantity) {
         this.price = price;
         this.quantity = quantity;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
